test: move small chunked decode tests to fetch-gzip.test.ts
fetch-retry-chunked.test.ts is scoped to the BUN-3BZF retry guard;
fetch-gzip.test.ts already hosts the chunked + Content-Encoding decode
coverage and has createNetServer / the chunked framing pattern.

Collapsed the four gzip sizes into one control case using incompressible
randomBytes so the compressed payload is actually multi-KiB rather than
a ~50 B gzipped run of 'x'.
