+++ test/js/web/fetch/fetch-retry-chunked.test.ts
@@ -4,10 +4,9 @@
 // body_out_str == None state itself is not deterministically reachable from
 // fetch().
 
-import { describe, expect, test } from "bun:test";
+import { expect, test } from "bun:test";
 import type { AddressInfo } from "node:net";
 import net from "node:net";
-import zlib from "node:zlib";
 
 // The server drops every third request without responding, so the client
 // that adopted the pooled socket observes on_close with response_stage ==
@@ -58,65 +57,5 @@ test("chunked uncompressed body over a retried keep-alive connection", async () 
   } finally {
     for (const s of sockets) s.destroy();
     await new Promise<void>(r => server.close(() => r()));
-  }
-});
-
-// The <=16 KiB chunked-body dispatcher routes identity bodies through the
-// append + decode-in-tail path and compressed bodies through the scratch
-// decode path. These pin both paths for a body that fits in one read.
-describe("small chunked body decode", () => {
-  function chunked(buf: Buffer): Buffer {
-    return Buffer.concat([Buffer.from(buf.length.toString(16) + "\r\n"), buf, Buffer.from("\r\n0\r\n\r\n")]);
-  }
-
-  async function serve(reply: Buffer, fn: (url: string) => Promise<void>): Promise<void> {
-    const sockets = new Set<net.Socket>();
-    const server = net.createServer(sock => {
-      sockets.add(sock);
-      sock.on("close", () => sockets.delete(sock));
-      sock.on("error", () => {});
-      sock.once("data", () => sock.write(reply));
-    });
-    await new Promise<void>((resolve, reject) => {
-      server.once("error", reject);
-      server.listen(0, "127.0.0.1", resolve);
-    });
-    const { port } = server.address() as AddressInfo;
-    try {
-      await fn(`http://127.0.0.1:${port}/`);
-    } finally {
-      for (const s of sockets) s.destroy();
-      await new Promise<void>(r => server.close(() => r()));
-    }
-  }
-
-  function reply(body: Buffer, extraHeader: string): Buffer {
-    return Buffer.concat([
-      Buffer.from(
-        "HTTP/1.1 200 OK\r\n" + "Transfer-Encoding: chunked\r\n" + "Connection: keep-alive\r\n" + extraHeader + "\r\n",
-      ),
-      chunked(body),
-    ]);
-  }
-
-  for (const n of [1, 256, 4096, 15 * 1024]) {
-    const body = Buffer.alloc(n, "x");
-    const gz = zlib.gzipSync(body, { level: 6 });
-
-    test.concurrent(`identity ${n}B`, async () => {
-      await serve(reply(body, ""), async url => {
-        const res = await fetch(url);
-        expect(res.status).toBe(200);
-        expect(Buffer.from(await res.arrayBuffer()).equals(body)).toBe(true);
-      });
-    });
-
-    test.concurrent(`gzip ${n}B`, async () => {
-      await serve(reply(gz, "Content-Encoding: gzip\r\n"), async url => {
-        const res = await fetch(url);
-        expect(res.status).toBe(200);
-        expect(Buffer.from(await res.arrayBuffer()).equals(body)).toBe(true);
-      });
-    });
   }
 });
